+++ lib/overcommit/hook/pre_commit/rails_schema_up_to_date.rb
@@ -7,9 +7,16 @@ module Overcommit::Hook::PreCommit
       elsif migration_files.none? && schema_files.any?
         return :fail, "You're trying to change the schema without adding a migration file"
       elsif migration_files.any? && schema_files.any?
-        latest_version = migration_files.map  { |file| file[/\d+/] }.sort.last
-        schema         = schema_files.map     { |file| File.read(file) }.join
-        up_to_date     = schema.include?(latest_version)
+        # Get the latest version from the migration filename. Use
+        # `File.basename` to prevent finding numbers that could appear in
+        # directories, such as the home directory of a user with a number in
+        # their username.
+        latest_version = migration_files.map do |file|
+          File.basename(file)[/\d+/]
+        end.sort.last
+
+        schema     = schema_files.map { |file| File.read(file) }.join
+        up_to_date = schema.include?(latest_version)
 
         unless up_to_date
           return :fail, "The latest migration version you're committing is " \